implemented child process calls
This means that we're technically fully functional now! \o/

Since we want to minimize/restore our window and don't want to
use any platform-/API-specific calls in app.cpp, a mechanism
to call back into main() had to be introduced. While doing that,
it made sense to switch the app's active flag and the "quit"
action over to the new mechanism as well.