don't eagerly try to execute files that are just data
On Linux, file systems with non-POSIX semantics like (ex)FAT and NTFS
are often mounted with options that cause *every* file to appear executable.
This caused GLBrowser to try to execute e.g. image files even though it's
supposed to run a viewer on them.

This change mitigates the issue by specifying for each file type association
whether it's reasonable to expect a file of that type to be executable.
If this is false (and it is for most file types), GLBrowser's default
action will be to run the primary associated application, regardless of
whether the file is marked as executable or not.